email notification
applying for a service will send email notification to the area manager in that area office.
fixed logout issues in express
added loading animation in express
added notification when payment is made and when documents are uploaded in express i.e email notification to the area manager in that area office.
added notification when documents are approved or rejected in optima i.e email notification to the client.
no application form fee and processing fee approval.
When application form fee is paid, the client can proceed to uploaded documents
The client can decide to pay for both processing fee and inspection fee at the same time.

To reduce improve and reduce
service application
steps so that client's
can apply for a service
faster without delay